+++ app/src/main/java/com/example/toucheeseapp/ui/screens/tab_Home/CartScreen.kt
@@ -11,7 +11,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -19,7 +18,7 @@ import com.example.toucheeseapp.data.model.carts_list.CartListResponseItem
 import com.example.toucheeseapp.data.model.carts_list.SelectAddOption
 import com.example.toucheeseapp.data.model.carts_optionChange.ChangedCartItem
 import com.example.toucheeseapp.data.token_manager.TokenManager
-import com.example.toucheeseapp.ui.components.CartTopAppBarComponent
+import com.example.toucheeseapp.ui.components.topbar.TopAppBarComponent
 import com.example.toucheeseapp.ui.components.ChangeOptionBottomSheetComponent
 import com.example.toucheeseapp.ui.components.CartItemComponent
 import com.example.toucheeseapp.ui.viewmodel.StudioViewModel
@@ -93,7 +92,7 @@ fun CartScreen(
     Scaffold(
         modifier = Modifier.safeDrawingPadding(),
         topBar = {
-            CartTopAppBarComponent(
+            TopAppBarComponent(
                 title = "장바구니",
                 showLeadingIcon = true,
                 showTrailingIcon = false,
